requirements.txt
pandas==2.3.0
numpy==1.26.4
sqlalchemy==2.0.30
requests==2.32.3
pyyaml==6.0.1
streamlit==1.35.0
plotly==5.22.0
pytest==8.2.2  # Only needed for running tests 